test: Restructure controlling tests into tests/unit/controlling/ package
Move all utils/controlling.py tests into a dedicated package for better
organization. Absorb scattered test files and deduplicate:

- test_control_trv.py: merged from test_boost_mode.py (safety overrides),
  test_race_condition_lock_coverage.py, test_grouped_trv_calibration.py,
  and test_unavailable_trv_no_operations.py (72 tests total)
- test_control_queue.py: moved (15 tests)
- test_helpers.py: merged with handle_window_open tests from
  test_window_no_off_mode.py (16 tests)
- test_control_cooler.py: moved (9 tests)
- test_task_manager.py: moved (5 tests)

Removed test_control_queue_retry.py (documentation-only pass stubs).
Kept test_window_no_off_mode.py for events/trv.py tests only.
