+++ tests/unit/controlling/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for utils/controlling.py — the TRV control orchestration layer."""
