+++ tests/unit/controlling/test_control_cooler.py
