+++ tests/unit/controlling/test_task_manager.py
